Make reading codelist files more restrictive
Restricts codelist files to endings ".xml". This allows
to make backups of codelist files that ends NOT with ".xml", i.e.
"codelist_8500.xml.20191216". The so named backups will not be taken
into account!

The former code read all files from the specified directories. Backups
resulted in duplicate codelists that lead to funny behaviors.